requirements.txt
fastapi==0.115.6
loguru==0.7.2
pydantic==2.10.3
python-dotenv==1.0.1
uvicorn-loguru-integration==0.3.0
